+++ io/js/src/main/scala/fs2/io/file/ReadCursorPlatform.scala
@@ -23,4 +23,4 @@ package fs2
 package io
 package file
 
-private[file] trait ReadCursorPlatform
+private[file] trait ReadCursorCompanionPlatform
