+++ io/js/src/main/scala/fs2/io/net/DatagramSocketGroupPlatform.scala
@@ -26,7 +26,6 @@ package net
 import cats.effect.kernel.Async
 import cats.effect.kernel.Resource
 import cats.syntax.all._
-import cats.effect.syntax.all._
 import com.comcast.ip4s.{Host, Port}
 import fs2.js.node.dgramMod
