+++ io/jvm/src/main/scala/fs2/io/net/SocketPlatform.scala
@@ -32,6 +32,10 @@ import java.net.InetSocketAddress
 import java.nio.channels.{AsynchronousSocketChannel, CompletionHandler}
 import java.nio.{Buffer, ByteBuffer}
 
+private[net] trait SocketPlatform[F[_]] {
+  def endOfInput: F[Unit]
+}
+
 private[net] trait SocketCompanionPlatform {
   private[net] def forAsync[F[_]: Async](
       ch: AsynchronousSocketChannel
